Move non-flight related utilities into proto-backplane-grpc. Rename proto-backplane-grpc-util proto-backplane-grpc-flight. Fixes #1186.
