+++ grpc-api-client/build.gradle
@@ -11,7 +11,7 @@ dependencies {
     compile project(':Base'),
             project(':DB'),
             project(':Util'),
-            project(':proto:proto-backplane-grpc-util'),
+            project(':proto:proto-backplane-grpc-flight'),
             project(':log-factory'),
             "com.google.flatbuffers:flatbuffers-java:1.12.0",
             "io.deephaven.barrage:barrage-format:0.2.0"
