+++ grpc-api/grpc-api.gradle
@@ -26,7 +26,7 @@ dependencies {
     compile project(':DB'),
             project(':Util'),
             project(':grpc-api-client'),
-            project(':proto:proto-backplane-grpc-util'),
+            project(':proto:proto-backplane-grpc-flight'),
             project(':open-api-lang-tools'),
             project(':log-factory'),
             project(':application-mode'),
