+++ java-client/flight/build.gradle
@@ -6,7 +6,7 @@ description = 'The Deephaven client flight library'
 
 dependencies {
     api project(':client-session')
-    implementation project(':proto:proto-backplane-grpc-util')
+    implementation project(':proto:proto-backplane-grpc-flight')
 
     Classpaths.inheritArrow(project, 'flight-core', 'api')
     Classpaths.inheritArrow(project, 'flight-grpc', 'implementation')
