+++ java-client/session/build.gradle
@@ -7,7 +7,6 @@ description = 'The Deephaven client session library'
 dependencies {
   api project(':qst')
   api project(':proto:proto-backplane-grpc')
-  implementation project(':proto:proto-backplane-grpc-util')
 
   annotationProcessor 'org.immutables:value:2.8.1'
   compileOnly project(':util-immutables')
