+++ proto/proto-backplane-grpc/src/main/java/io/deephaven/grpc_api/util/Exceptions.java
