+++ proto/proto-backplane-grpc/src/main/java/io/deephaven/grpc_api/util/ExportTicketHelper.java
@@ -8,7 +8,6 @@ import com.google.protobuf.ByteStringAccess;
 import com.google.rpc.Code;
 import io.deephaven.proto.backplane.grpc.TableReference;
 import io.deephaven.proto.backplane.grpc.Ticket;
-import org.apache.arrow.flight.impl.Flight;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
@@ -19,18 +18,7 @@ public class ExportTicketHelper {
     public static final String FLIGHT_DESCRIPTOR_ROUTE = "export";
 
     /**
-     * Convenience method to convert from export id to {@link Flight.Ticket}.
-     *
-     * @param exportId the export id
-     * @return a grpc Ticket wrapping the export id
-     */
-    public static Flight.Ticket exportIdToArrowTicket(int exportId) {
-        final byte[] dest = exportIdToBytes(exportId);
-        return Flight.Ticket.newBuilder().setTicket(ByteStringAccess.wrap(dest)).build();
-    }
-
-    /**
-     * Convenience method to convert from export id to {@link Flight.Ticket}.
+     * Convenience method to convert from export id to {@link Ticket}.
      *
      * @param exportId the export id
      * @return a grpc Ticket wrapping the export id
@@ -41,19 +29,7 @@ public class ExportTicketHelper {
     }
 
     /**
-     * Convenience method to convert from export id to {@link Flight.FlightDescriptor}.
-     *
-     * @param exportId the export id
-     * @return a grpc Ticket wrapping the export id
-     */
-    public static Flight.FlightDescriptor exportIdToDescriptor(int exportId) {
-        return Flight.FlightDescriptor.newBuilder()
-                .setType(Flight.FlightDescriptor.DescriptorType.PATH).addPath(FLIGHT_DESCRIPTOR_ROUTE)
-                .addPath(Integer.toString(exportId)).build();
-    }
-
-    /**
-     * Convenience method to convert from {@link Flight.Ticket} to export id.
+     * Convenience method to convert from {@link Ticket} to export id.
      *
      * <p>
      * Ticket's byte[0] must be {@link ExportTicketHelper#TICKET_PREFIX}, bytes[1-4] are a signed int export id in
@@ -64,22 +40,6 @@ public class ExportTicketHelper {
      * @return the export id that the Ticket wraps
      */
     public static int ticketToExportId(final Ticket ticket, final String logId) {
-        return ticketToExportIdInternal(
-                ticket.getTicket().asReadOnlyByteBuffer().order(ByteOrder.LITTLE_ENDIAN), logId);
-    }
-
-    /**
-     * Convenience method to convert from {@link Flight.Ticket} to export id.
-     *
-     * <p>
-     * Ticket's byte[0] must be {@link ExportTicketHelper#TICKET_PREFIX}, bytes[1-4] are a signed int export id in
-     * little-endian.
-     *
-     * @param ticket the grpc Ticket
-     * @param logId an end-user friendly identification of the ticket should an error occur
-     * @return the export id that the Ticket wraps
-     */
-    public static int ticketToExportId(final Flight.Ticket ticket, final String logId) {
         return ticketToExportIdInternal(
                 ticket.getTicket().asReadOnlyByteBuffer().order(ByteOrder.LITTLE_ENDIAN), logId);
     }
@@ -108,7 +68,6 @@ public class ExportTicketHelper {
                 : ticketToExportIdInternal(ticket.asReadOnlyBuffer().order(ByteOrder.LITTLE_ENDIAN), logId);
     }
 
-
     /**
      * Convenience method to convert from {@link ByteBuffer} to export ticket. Most efficient when {@code ticket} is
      * {@link ByteOrder#LITTLE_ENDIAN}.
@@ -127,84 +86,6 @@ public class ExportTicketHelper {
         final ByteBuffer lebb = ticket.order() == ByteOrder.LITTLE_ENDIAN ? ticket
                 : ticket.asReadOnlyBuffer().order(ByteOrder.LITTLE_ENDIAN);
         return Ticket.newBuilder().setTicket(ByteStringAccess.wrap(lebb)).build();
-    }
-
-    /**
-     * Convenience method to convert from {@link Flight.FlightDescriptor} to export id.
-     *
-     * <p>
-     * Descriptor must be a path.
-     *
-     * @param descriptor the grpc Ticket
-     * @param logId an end-user friendly identification of the ticket should an error occur
-     * @return the export id that the Ticket wraps
-     */
-    public static int descriptorToExportId(final Flight.FlightDescriptor descriptor, final String logId) {
-        if (descriptor == null) {
-            throw Exceptions.statusRuntimeException(Code.FAILED_PRECONDITION,
-                    "Could not resolve descriptor '" + logId + "': is empty");
-        }
-        if (descriptor.getType() != Flight.FlightDescriptor.DescriptorType.PATH) {
-            throw Exceptions.statusRuntimeException(Code.FAILED_PRECONDITION,
-                    "Could not resolve descriptor '" + logId + "': not a path");
-        }
-        if (descriptor.getPathCount() != 2) {
-            throw Exceptions.statusRuntimeException(Code.FAILED_PRECONDITION,
-                    "Could not resolve descriptor '" + logId + "': unexpected path length (found: "
-                            + TicketRouterHelper.getLogNameFor(descriptor) + ", expected: 2)");
-        }
-
-        try {
-            return Integer.parseInt(descriptor.getPath(1));
-        } catch (final NumberFormatException nfe) {
-            throw Exceptions.statusRuntimeException(Code.FAILED_PRECONDITION,
-                    "Could not resolve descriptor '" + logId + "': export id not numeric (found: "
-                            + TicketRouterHelper.getLogNameFor(descriptor) + ")");
-        }
-    }
-
-    /**
-     * Convenience method to convert from a Flight.Ticket to a Flight.FlightDescriptor.
-     *
-     * @param ticket the ticket to convert
-     * @param logId an end-user friendly identification of the ticket should an error occur
-     * @return a flight descriptor that represents the ticket
-     */
-    public static Flight.FlightDescriptor ticketToDescriptor(final Ticket ticket, final String logId) {
-        return exportIdToDescriptor(ticketToExportId(ticket, logId));
-    }
-
-    /**
-     * Convenience method to convert from a Flight.Ticket to a Flight.FlightDescriptor.
-     *
-     * @param ticket the ticket to convert
-     * @param logId an end-user friendly identification of the ticket should an error occur
-     * @return a flight descriptor that represents the ticket
-     */
-    public static Flight.FlightDescriptor ticketToDescriptor(final Flight.Ticket ticket, final String logId) {
-        return exportIdToDescriptor(ticketToExportId(ticket, logId));
-    }
-
-    /**
-     * Convenience method to convert from a Flight.Descriptor to a Flight.Ticket.
-     *
-     * @param descriptor the descriptor to convert
-     * @param logId an end-user friendly identification of the ticket should an error occur
-     * @return a flight ticket that represents the descriptor
-     */
-    public static Flight.Ticket descriptorToArrowTicket(final Flight.FlightDescriptor descriptor, final String logId) {
-        return exportIdToArrowTicket(descriptorToExportId(descriptor, logId));
-    }
-
-    /**
-     * Convenience method to convert from a Flight.Descriptor to a Flight.Ticket.
-     *
-     * @param descriptor the descriptor to convert
-     * @param logId an end-user friendly identification of the ticket should an error occur
-     * @return a flight ticket that represents the descriptor
-     */
-    public static Ticket descriptorToTicket(final Flight.FlightDescriptor descriptor, final String logId) {
-        return wrapExportIdInTicket(descriptorToExportId(descriptor, logId));
     }
 
     /**
@@ -254,7 +135,7 @@ public class ExportTicketHelper {
         return FLIGHT_DESCRIPTOR_ROUTE + "/" + ticketToExportId(ticket, logId);
     }
 
-    private static byte[] exportIdToBytes(int exportId) {
+    static byte[] exportIdToBytes(int exportId) {
         final byte[] dest = new byte[5];
         dest[0] = TICKET_PREFIX;
         dest[1] = (byte) exportId;
@@ -264,7 +145,7 @@ public class ExportTicketHelper {
         return dest;
     }
 
-    private static String byteBufToHex(final ByteBuffer ticket) {
+    static String byteBufToHex(final ByteBuffer ticket) {
         StringBuilder sb = new StringBuilder();
         for (int i = ticket.position(); i < ticket.limit(); ++i) {
             sb.append(String.format("%02x", ticket.get(i)));
@@ -272,7 +153,7 @@ public class ExportTicketHelper {
         return sb.toString();
     }
 
-    private static int ticketToExportIdInternal(final ByteBuffer ticket, final String logId) {
+    static int ticketToExportIdInternal(final ByteBuffer ticket, final String logId) {
         if (ticket.order() != ByteOrder.LITTLE_ENDIAN) {
             throw Exceptions.statusRuntimeException(Code.FAILED_PRECONDITION,
                     "Could not resolve ticket '" + logId + "': ticket is not in LITTLE_ENDIAN order");
