+++ proto/proto-backplane-grpc/src/test/java/io/deephaven/grpc_api/util/OperationHelperTest.java
