+++ settings.gradle
@@ -54,7 +54,7 @@ pyMods.each {
 include 'fishconfig-local'
 include 'DHProcess'
 include 'proto:proto-backplane-grpc'
-include 'proto:proto-backplane-grpc-util'
+include 'proto:proto-backplane-grpc-flight'
 include 'proto:raw-js-openapi'
 include 'protoc'
